XW-4232 | get rid of dom-helper
